+++ backend/app/controllers/verify.js
@@ -1,0 +1,18 @@
+var jwt = require('jsonwebtoken');
+var SECRET = require('../../SECRET/config').secret;
+module.exports = (req, res, next) => {
+    var token = req.body.token || req.query.token || req.headers['x-access-token'];
+    if (!token) {
+        res.status(401).json({'error':'Token required'});
+        return;
+    } else {
+        jwt.verify(token, SECRET, (err, decoded) => {
+            if (err) {
+                return res.status(500).json({'error': 'Failed to Authenticate Token'});
+            } else {
+                req.decoded = decoded;
+                next();
+            }
+        });
+    }
+};
